# project search: Increase perf up to 10x by batching `git status` calls