--- a/crates/worktree/Cargo.toml
+++ b/crates/worktree/Cargo.toml
@@ -41,6 +41,7 @@ schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
+smallvec.workspace = true
 smol.workspace = true
 sum_tree.workspace = true
 text.workspace = true