@mrnugget 2024-08-27
Merging this, even though it seems like perf on `main` has decreased after https://github.com/zed-industries/zed/pull/16915